backend/requirements.txt
fastapi
uvicorn[standard]
boto3
python-docx
pymupdf
python-multipart
python-dotenv